Add Schema-Creation to createSchema function
Currently when a tableName with a dot is presented the name can correctly
be split up into schema and table name. But when the schema does not
exist the whole proces breaks as the table can not be created in a
non-existing space.

This commit adds an SQL-statement to the createSchema-function that – in
case a schema was given – will create the schema when it does not yet
exist.